% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accuracy.R
\name{accuracy}
\alias{accuracy}
\title{accuracy}
\usage{
accuracy(pi, mu, Sigma, index)
}
\description{
Compute accuracy for the selected dimensions (index) for a specific component comp0,in normal mixture model.
}
