% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centroid_sphBregman_GMM.R
\name{centroid_sphBregman_GMM}
\alias{centroid_sphBregman_GMM}
\title{The algorithmic prototype of Wasserstein Barycenter using Bregman ADMM}
\usage{
centroid_sphBregman_GMM(stride, instanceW, supp, w, c0, options)
}
\arguments{
\item{instanceW}{the weights of each sample}

\item{supp}{mu and sigma informations in GMM models}

\item{w}{the weighhts of each cluster}

\item{c0}{the initialization}

\item{options}{MLMoptions}
}
\value{
the final GMM result
}
\description{
The algorithmic prototype of Wasserstein Barycenter using Bregman ADMM
}
