% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_MLM.R
\name{em_MLM}
\alias{em_MLM}
\title{em_MLM}
\usage{
em_MLM(X, Xlogit, Y, cinit, betainit, MLMoption)
}
\arguments{
\item{X}{all non-dummy features that are used in the logistic model}

\item{Xlogit}{include the dummy variables}

\item{Y}{the response variable}

\item{cinit}{some information about the Gaussian model}

\item{betainit}{the coefficients in logistics model}

\item{MLMoption}{some necessary variabls}
}
\value{
the optimized parameters
}
\description{
Use EM algorithm to optimize some parameters.
}
