% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_MLM_maw.R
\name{em_MLM_maw}
\alias{em_MLM_maw}
\title{em_MLM_maw}
\usage{
em_MLM_maw(X, Xlogit, Y, cinit, betainit, MLMoption, cprior)
}
\arguments{
\item{X}{all non-dummy features that are used in the logistic model}

\item{Xlogit}{include the dummy variables}

\item{Y}{the response variable}

\item{cinit}{some information about the Gaussian model}

\item{betainit}{the coefficients in logistics model}

\item{MLMoption}{some necessary variabls}

\item{cprior}{the given template GMM model}
}
\value{
the optimized parameters
}
\description{
Use maw distance to advance the iteration in em_MLM.
}
