% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{estimate}
\alias{estimate}
\title{estimate}
\usage{
estimate(X, Xlogit, Y, MLMoption)
}
\arguments{
\item{X}{all non-dummy feature variables}

\item{Xlogit}{all feature variables}

\item{Y}{the response variable}

\item{MLMoption}{all necessary variables}
}
\value{
all informations about the gmm and logistic model
}
\description{
use EM algorithm for estimating MLM.
}
