% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateBestSD.R
\name{estimateBestSD}
\alias{estimateBestSD}
\title{estimate the model that have the max AUC}
\usage{
estimateBestSD(X, Xlogit, Y, MLMoption, seedlist)
}
\arguments{
\item{X}{all non-dummy feature variables}

\item{Xlogit}{all feature variables}

\item{Y}{the response variable}

\item{MLMoption}{all necessary variables}

\item{seedlist}{the list of random seeds}
}
\value{
all informations about the gmm and logistic model
}
\description{
Repeat EM algorithm for estimating MLM using multiple random seeds for kmeans initialization. Choose the best seed based on training accuracy or AUC.
}
