% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalModel.R
\name{finalModel}
\alias{finalModel}
\title{final Model}
\usage{
finalModel(
  cvAUCfinal,
  ncmp,
  nseeds,
  rangeSeed,
  vargmm,
  vlasso,
  Y,
  Xs,
  X,
  Indi,
  varreg = NULL,
  alphaLasso = 0.8,
  stopratio = 1e-05,
  verbose = 1,
  constrain = "DIAS",
  diagshrink = 0.9,
  kappa = -1,
  AUC = 1,
  DISTR = "binomial"
)
}
\arguments{
\item{cvAUCfinal}{the AUC results in cross-verified}

\item{ncmp}{the number of clusters}

\item{nseeds}{the number of random seeds}

\item{rangeSeed}{the largest number of random seeds}

\item{vargmm}{the index of variables that are used in gmm model}

\item{vlasso}{Lambda for Lasso penalty}

\item{Y}{the response variable}

\item{Xs}{the standardized independent variables}

\item{X}{the raw input of independent variables}

\item{Indi}{the dummy variables}

\item{varreg}{(optional) Variables for logistic regression. If NULL, uses all variables}

\item{alphaLasso}{Elastic net parameter (default 0.8)}

\item{stopratio}{Convergence threshold (default 1.0e-5)}

\item{verbose}{Verbosity flag (default 1)}

\item{constrain}{Covariance constraint (default "DIAS")}

\item{diagshrink}{Diagonal shrinkage (default 0.9)}

\item{kappa}{Instance weighting (default -1)}

\item{AUC}{Use AUC for selection (default 1)}

\item{DISTR}{Distribution family (default "binomial")}
}
\value{
a list of clustering results
}
\description{
use AUC to decide the model
}
