% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GaussWasserstein.R
\name{GaussWasserstein}
\alias{GaussWasserstein}
\title{Compute the pairwise squared Wasserstein distance between two GMM model}
\usage{
GaussWasserstein(dim, supp1, supp2)
}
\arguments{
\item{dim}{the dimension of the data}

\item{supp1}{mu and sigma in model1}

\item{supp2}{mu and sigma in model2}
}
\value{
The distance matrix
}
\description{
Compute the pairwise squared Wasserstein distance between two GMM model
}
