% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmIntercept.R
\name{glmIntercept}
\alias{glmIntercept}
\title{glm Intercept}
\usage{
glmIntercept(Y, wt, dim, DISTR)
}
\arguments{
\item{Y}{the respinse variable}

\item{wt}{the weight}

\item{dim}{the number of features}

\item{DISTR}{default = binominal}
}
\value{
the intercept coefficient of the elastic net regression
}
\description{
Fit a linear model with only intercept term. This is used when the number of data points for estimation is extremely small.
}
