% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMMFormatConvert.R
\name{GMMFormatConvert}
\alias{GMMFormatConvert}
\title{GMM format convert}
\usage{
GMMFormatConvert(dim, c)
}
\arguments{
\item{dim}{dimensions of Gaussian distribution for each cluster}

\item{c}{a list that store the weight, mean and sigma of each cluster}
}
\value{
a list of informations of each cluster
}
\description{
convert the compressed variables to the Gaussian mixtured model format.
}
