% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initemMLM.R
\name{initemMLM}
\alias{initemMLM}
\title{init emMLM}
\usage{
initemMLM(X, Xlogit, Y, MLMoption)
}
\arguments{
\item{X}{all non-dummy features that are used in logistic model}

\item{Xlogit}{include all features and dummy feature}

\item{Y}{the response variable}

\item{MLMoption}{some necessary variables}
}
\value{
a list of initialization informations of each cluster
}
\description{
use EM algorithm to initiate a model.
}
