% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/km.R
\name{km}
\alias{km}
\title{km}
\usage{
km(dat, nkm, thred, kmseed)
}
\arguments{
\item{dat}{the input data with every sample stored in one row}

\item{nkm}{the number of clusters prespecified}

\item{thred}{the threshold for determining when to stop the iteration in kmeans. Suggested value for thred: 1.0e-4 to 1.0e-6}

\item{kmseed}{a random number seed that specifies a way to assign random numbers to each sample}
}
\value{
the center of each cluster and the identity of each sample
}
\description{
use k-means algorithm to initiate the Gussian mixture model
}
