% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KMclassify.R
\name{KMclassify}
\alias{KMclassify}
\title{classify based on a given MLM model}
\usage{
KMclassify(a, mu, sigma, beta, X, Xlogit)
}
\value{
classification
}
\description{
classify based on a given MLM model
}
