% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmRmEmpty.R
\name{kmRmEmpty}
\alias{kmRmEmpty}
\title{kmRmEmpty}
\usage{
kmRmEmpty(dat, nkm, cdbkinit)
}
\arguments{
\item{dat}{the input data with every sample stored in one row}

\item{nkm}{the number of clusters prespecified}

\item{cdbkinit}{the randomly initiated centers of each cluster}
}
\value{
the center of each cluster and the identity of each sample
}
\description{
to remove the empty clusters of k-means.
}
