% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logisticIntercept.R
\name{logisticIntercept}
\alias{logisticIntercept}
\title{Compute the logistic model's intercept}
\usage{
logisticIntercept(Y, wt, dim)
}
\arguments{
\item{Y}{the response variable}

\item{wt}{weights}

\item{dim}{the dimension of dataset(number of features)}
}
\value{
the beta coefficients
}
\description{
Compute the logistic model's intercept
}
