% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoglikelihoodYBalance.R
\name{LoglikehoodYBalance}
\alias{LoglikehoodYBalance}
\title{LoglikehoodYBalance}
\usage{
LoglikehoodYBalance(pyi, Y, w)
}
\arguments{
\item{pyi}{P(Y=1|X,Z)}

\item{Y}{the resonse variables}

\item{w}{the weigh}
}
\value{
the loglike Y
}
\description{
Compute the conditional log likelihood of Y given X.As the classes may be very unbalanced, adjust the likelihood by assuming the class proportions are given in input w. If w is set to the empirical proportions of the classes, the output is simply the conditional log likelihood of the observed Y.
}
