% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLMclassify.R
\name{MLMclassify}
\alias{MLMclassify}
\title{MLM classify}
\usage{
MLMclassify(a, mu, sigma, beta, X, Xlogit)
}
\arguments{
\item{a}{the weight of each cluster}

\item{mu}{the mean of each cluster}

\item{sigma}{the sigma matrix of each cluster}

\item{beta}{the logistic coefficients of each cluster}

\item{X}{all non-dummy feature variables}

\item{Xlogit}{all feature variables}
}
\value{
P(Y=1|given X) and P(Z=c cluster|given X)
}
\description{
classify based on a given MLM model
}
