% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MstepObj.R
\name{MstepObj}
\alias{MstepObj}
\title{Compute the Q(theta'|theta) function in the EM algorithm}
\usage{
MstepObj(dim, numdata, supp, ww, X, pij)
}
\arguments{
\item{dim}{the number of features}

\item{numdata}{the number of samples}

\item{supp}{mu and sigma informations about the GMM model}

\item{ww}{the weights of old and new model}

\item{X}{the features of each sample}

\item{pij}{the posterior matrix pij should be pre-computed based on previous model, not the updated model}
}
\value{
the likelihood
}
\description{
Compute the Q(theta'|theta) function in the EM algorithm
}
