% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvnormpdf.R
\name{mvnormpdf}
\alias{mvnormpdf}
\title{mvnormpdf}
\usage{
mvnormpdf(x, mu, Sigma)
}
\arguments{
\item{x}{p.n array of n values of p-dim MV normal}

\item{mu}{column p vector mean}

\item{Sigma}{p.p variance matrix}
}
\value{
the value of pdf
}
\description{
mvnormpdf
}
