% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_beta_heatmap.R
\name{plot_beta_heatmap}
\alias{plot_beta_heatmap}
\title{Plot the heatmap of the beta of the best cluster number}
\usage{
plot_beta_heatmap(
  beta_result,
  output_file = NULL,
  width = 5.5,
  height = 5,
  res = 300
)
}
\arguments{
\item{beta_result}{the regression beta coefficient of each cluster}

\item{output_file}{If you just want to have a look, let it be NULL.}

\item{width}{default}

\item{height}{default}

\item{res}{default}
}
\value{
a heatmap
}
\description{
Plot the heatmap of the beta of the best cluster number
}
\examples{
plot_beta_heatmap(beta_result), plot_beta_heatmap(beta_result, output_file = "beta_heatmap.png")
}
