% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{predict_class}
\alias{predict_class}
\title{Predict Class Labels (expects newx standardized same as training Xs)}
\usage{
predict_class(fit, newx, Indi = NULL, threshold = 0.5)
}
\description{
Predict Class Labels (expects newx standardized same as training Xs)
}
