% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_data}
\alias{read_data}
\title{Read dataset and prepare X/Xs/Y/Indi (silent & minimal return)}
\usage{
read_data(dat_path, ycol = NULL, Indi_col = NULL, encoding = NULL)
}
\arguments{
\item{dat_path}{Path to data file.}

\item{ycol}{Optional Y column (index or name). If NULL, use last column.}

\item{Indi_col}{Optional indicator column (single index or single name).}

\item{encoding}{Optional file encoding for text files.}
}
\value{
list(dim, numdata, rawdat, X, Xs, Y, Indi)
}
\description{
Read dataset and prepare X/Xs/Y/Indi (silent & minimal return)
}
