% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqrtm_eig.R
\name{sqrtm_eig}
\alias{sqrtm_eig}
\title{Use eigen value decomposition to solve squre root}
\usage{
sqrtm_eig(A)
}
\arguments{
\item{A}{matrix}
}
\value{
the square root matrix
}
\description{
Use eigen value decomposition to solve squre root
}
