% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VS.R
\name{VS}
\alias{VS}
\title{VS}
\usage{
VS(comp, a2, mu2, sigma2, dimgmm)
}
\arguments{
\item{comp}{use vs on which components}

\item{a2}{the weight of each cluster}

\item{mu2}{the mean of each cluster}

\item{sigma2}{the covariance matrix of each cluster}

\item{dimgmm}{dimensions of gmm}
}
\value{
a list of information about the explainable variables
}
\description{
variable selection for Gaussian mixtured model.
}
