% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checksingular.R
\name{checksingular}
\alias{checksingular}
\title{check singular}
\usage{
checksingular(sigma, Xvar, shrinkrate)
}
\arguments{
\item{sigma}{the sigma matrix}

\item{Xvar}{the gmm features variables}

\item{shrinkrate}{the rate of shrinking}
}
\value{
the new singular sigma matrix
}
\description{
Check whether covariance matrix is ill-conditioned or nearly singular. If so, regularize covariance by changing the diagonal elements.
}
