% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalModel.R
\name{finalModel}
\alias{finalModel}
\title{final Model}
\usage{
finalModel(
  cvAUCfinal,
  ncmp,
  nseeds,
  rangeSeed,
  vargmm,
  Y,
  Xs,
  X,
  Indi,
  MLMoption
)
}
\arguments{
\item{cvAUCfinal}{the AUC results in cross-verified}

\item{ncmp}{the number of clusters}

\item{nseeds}{the number of random seeds}

\item{rangeSeed}{the largest number of random seeds}

\item{vargmm}{the index of variables that are used in gmm model}

\item{Y}{the response variable}

\item{Xs}{the standardized independent variables}

\item{Indi}{the dummy variables}

\item{MLMoption}{all necessary variables}
}
\value{
a list of clustering results
}
\description{
use AUC to decide the model
}
