% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{fit_model}
\alias{fit_model}
\title{Quick GeMLR Fit Without Cross-Validation}
\usage{
fit_model(
  X,
  Y,
  Indi = NULL,
  K = 2,
  vargmm = NULL,
  nseeds = 3,
  rangeSeed = 500,
  vlasso = NULL,
  alphaLasso = 0.5,
  verbose = 0
)
}
\arguments{
\item{Indi}{Optional indicator (dummy) matrix/data.frame (n x q), appended to logistic design.}

\item{K}{Number of clusters (default 2).}

\item{vargmm}{Column names or indices to use for GMM. Use compute_gemlr_params() to obtain.}

\item{nseeds}{Number of kmeans seeds to try (default 3).}

\item{rangeSeed}{Upper bound for random seed sampling (default 500).}

\item{vlasso}{Lambda for Lasso penalty. Use compute_gemlr_params() to obtain.}

\item{alphaLasso}{Elastic-net alpha for glmnet (default 0.5).}

\item{verbose}{Verbosity flag passed to GeMLR internals (0 = silent).}

\item{x}{Numeric matrix/data.frame of predictors (n x p), ALREADY standardized (Xs from read_data()).}

\item{y}{Binary response vector (0/1), length n.}
}
\value{
An object of class "GeMLR_fit".
}
\description{
Quick GeMLR Fit Without Cross-Validation
}
