% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_MLMoption.R
\name{init_MLMoption}
\alias{init_MLMoption}
\title{init MLMoption parameters}
\usage{
init_MLMoption(
  alphaLasso = 0.5,
  vlasso = 1,
  numcmp = 2,
  stopratio = 1e-05,
  verbose = 1,
  minloop = 3,
  maxloop = 50,
  constrain = "DIAS",
  diagshrink = 0.9,
  kmseed = 0,
  algorithm = 1,
  AUC = 1,
  DISTR = "binomial",
  NOEM = 0,
  Yalpha = 1
)
}
\value{
a list of necessary parameters in a list named 'MLMoption'
}
\description{
Specify the GeM-LR model via MLMoption
}
