% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_data}
\alias{read_data}
\title{read dataset}
\usage{
read_data(
  dat_road,
  sep_mark = " ",
  num_gmm = NULL,
  alphaLasso = 0.8,
  ycol = NULL,
  Indi_col = 1,
  gmm_var = NULL
)
}
\arguments{
\item{dat_road}{the road of your .txt dataset file}

\item{num_gmm}{the number of variables that are used in the gmm model, default=5}

\item{alphaLasso}{the alpha number that are used in the elastic net regression}
}
\value{
a list of dataset and necessary variables
}
\description{
read the dataset and relevant parameters.
}
