% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_data}
\alias{read_data}
\title{read dataset}
\usage{
read_data(
  dat_road,
  num_gmm = 5,
  ncmp = c(2, 3, 4),
  alphaLasso = 0.8,
  nseeds = 20,
  rangeSeed = 30,
  kkk = 5
)
}
\arguments{
\item{dat_road}{the road of your .txt dataset file}

\item{num_gmm}{the number of variables that are used in the gmm model, default=5}

\item{ncmp}{a vector that store the different number of clusters}

\item{alphaLasso}{the alpha number that are used in the elastic net regression}

\item{nseeds}{the number of random number seeds}

\item{rangeSeed}{the largest number in seeds' list}

\item{kkk}{the number of folds of cross-validation}
}
\value{
a list of dataset and necessary variables
}
\description{
read the dataset and relevant parameters.
}
