% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCV.R
\name{runCV}
\alias{runCV}
\title{run CV}
\usage{
runCV(
  k = 5,
  ncmp = c(2, 3, 4),
  nseeds = 20,
  rangeSeed = 30,
  vargmm,
  Y,
  X,
  Indi,
  MLMoption
)
}
\arguments{
\item{k}{the number of folds of cross-validation}

\item{ncmp}{the number of clusters}

\item{nseeds}{the number of random seeds}

\item{rangeSeed}{the largest number among random seeds}

\item{vargmm}{the index of variables that are used in gmm}

\item{Y}{the response variable}

\item{X}{all independent variables}

\item{Indi}{dummy variable}

\item{MLMoption}{all necessary variables}
}
\value{
all AUC informations and the final classification result
}
\description{
Run Cross-Verified for GeM-LR with number of components being ncmp.
}
