% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCV.R
\name{runCV}
\alias{runCV}
\title{run CV}
\usage{
runCV(kkk, ncmp, nseeds, rangeSeed, vargmm, Y1, X1, Indi, MLMoption)
}
\arguments{
\item{kkk}{the number of folds of cross-validation}

\item{ncmp}{the number of clusters}

\item{nseeds}{the number of random seeds}

\item{rangeSeed}{the largest number among random seeds}

\item{vargmm}{the number of variables that are used in gmm}

\item{Y1}{the response variable}

\item{X1}{all non-dummy variables}

\item{Indi}{dummy variable}

\item{MLMoption}{all necessary variables}
}
\value{
all AUC informations and the final classification result
}
\description{
Run Cross-Verified for GeM-LR with number of components being ncmp.
}
